change data type of usn to biginteger in table usn_lists

+++ database/migrations/2023_06_14_023954_create_usn_lists_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('usn_lists', function (Blueprint $table) {
-            $table->integer('usn');
+            $table->bigInteger('usn');
             $table->string('fname');
             $table->string('lname');
         });
